Move callbacks files from collections to server/callbacks
