+++ packages/lesswrong/server/callbacks/messageCallbacks.ts
@@ -1,6 +1,6 @@
-import { addEditableCallbacks } from '../../../server/editor/make_editable_callbacks.js'
-import Messages, { makeEditableOptions } from './collection.js'
-import Conversations from '../conversations/collection'
+import { addEditableCallbacks } from '../editor/make_editable_callbacks.js'
+import Messages, { makeEditableOptions } from '../../lib/collections/messages/collection'
+import Conversations from '../../lib/collections/conversations/collection'
 import { addCallback } from 'meteor/vulcan:core'
 
 addEditableCallbacks({collection: Messages, options: makeEditableOptions})
